# docs/swagger.yaml
basePath: /v1
definitions:
  model.APIError:
    properties:
      error_message:
        type: string
      error_status:
        type: integer
    type: object
...
  model.Group:
    properties:
      bootdisk:
        type: string
      bootmethod:
        type: string
      callbackurl:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      device:
        type: string
      dns:
        type: string
      gateway:
        type: string
      host:
        items:
          $ref: '#/definitions/model.Host'
        type: array
      id:
        type: integer
      image_id:
        type: integer
      ks:
        type: string
      name:
        type: string
      netmask:
        type: string
      ntp:
        type: string
      options:
        additionalProperties:
          type: string
        type: object
      syslog:
        type: string
      updated_at:
        type: string
      vlan:
        type: string
    type: object
...
  model.Host:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      domain:
        type: string
      expires_at:
        type: string
      first_seen:
        type: string
      group:
        $ref: '#/definitions/model.Group'
      group_id:
        type: integer
      hostname:
        type: string
      id:
        type: integer
      ip:
        type: string
      ks:
        type: string
      last_seen:
        type: string
      last_seen_relay:
        description: DHCP parameters
        type: string
      mac:
        type: string
      missing_options:
        type: string
      progress:
        type: integer
      progresstext:
        type: string
      reimage:
        type: boolean
      updated_at:
        type: string
    type: object
  model.HostForm:
    properties:
      domain:
        type: string
      group_id:
        type: integer
      hostname:
        type: string
      ip:
        type: string
      ks:
        type: string
      mac:
        type: string
      progress:
        type: integer
      progresstext:
        type: string
      reimage:
        type: boolean
    type: object
  model.Image:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      hash:
        type: string
      id:
        type: integer
      iso_image:
        type: string
      path:
        type: string
      size:
        type: integer
      updated_at:
        type: string
    type: object
  model.ImageForm:
    properties:
      description:
        type: string
      hash:
        type: string
      iso_image:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
...
info:
  contact: {}
  description: VMware Imaging Appliances written in GO with full HTTP-REST API
  title: go-via
  version: "0.1"
paths:
...
  /hosts:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Host'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Get all hosts
      tags:
      - hosts
    post:
      consumes:
      - application/json
      parameters:
      - description: Add a host
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.HostForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Host'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Create a new host
      tags:
      - hosts
  /hosts/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Remove an existing host
      tags:
      - hosts
    get:
      consumes:
      - application/json
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Host'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Get an existing Host
      tags:
      - hosts
    patch:
      consumes:
      - application/json
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update a host
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.HostForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Host'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Update an existing host
      tags:
      - hosts
  /hosts/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: Fields to search for
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.Host'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Host'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Search for an host
      tags:
      - hosts
  /images:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Image'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Get all images
      tags:
      - images
    post:
      consumes:
      - application/json
      parameters:
      - description: Add image
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.ImageForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Image'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Create a new image
      tags:
      - images
  /images/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Remove an existing image
      tags:
      - images
    get:
      consumes:
      - application/json
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Image'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Get an existing image
      tags:
      - images
    patch:
      consumes:
      - application/json
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update an image
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.ImageForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Image'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      summary: Update an existing image
      tags:
      - images
...
swagger: "2.0"